fix(transforms): store double, not float; document host_io wrapper rationale
Ruling 1: switch TransformStore's Transform to double pos[3]/rot[9] and
every public-API float parameter/std::array<float,N> to double, matching
Python's float (a C double) bit-for-bit. Task 4 routes every object's
transform through this store, so truncating to float on every write would
make motion integration drift from today's SDK-visible behaviour — the
spec's "byte-identical" goal outranks the smaller POD. Restored
test_rotation_round_trips to exact equality (no pytest.approx) now that
both backends agree bit-for-bit; that agreement is the whole point of the
shared conformance suite. Renderer float32 downconversion stays out of
scope for Task 6.

Ruling 2: documented why engine/host_io.py carries nine transform_* wrapper
functions that engine.appc.transform_store.NativeTransformStore itself
never calls — they exist solely so the binding manifest can validate the
transform surface and fail loudly on a stale build, while the hot-path
store bypasses this façade and calls _dauntless_host directly.

+++ native/src/transforms/include/dauntless/transform_store.h
@@ -18,8 +18,15 @@ public:
 //
 // Position and rotation live here rather than on the Python object so that
 // the renderer and the motion integrator can read them without crossing the
-// language boundary. Rotation is row-major nine floats, matching TGMatrix3's
+// language boundary. Rotation is row-major nine doubles, matching TGMatrix3's
 // mIJ = row I, column J. Column-vector convention: column 1 is forward.
+//
+// Storage is double, not float: Python's `float` IS a C double, and this store
+// is the sole owner of every object's transform (Task 4). Truncating to float
+// on every write would make the native backend diverge from the pure-Python
+// one over an integration run, breaking the "byte-identical from the SDK's
+// point of view" goal. The renderer downconverts to float at the GL boundary
+// (Task 6), not here.
 //
 // Slots are generation-counted: a handle to a freed slot fails loudly rather
 // than silently reading whatever object recycled its index. Indices stay
@@ -32,8 +39,8 @@ public:
 class TransformStore {
 public:
     struct Transform {
-        float pos[3];
-        float rot[9];
+        double pos[3];
+        double rot[9];
     };
 
     // Returns (index, generation). New slots are identity at the origin.
@@ -42,19 +49,19 @@ public:
 
     bool valid(std::uint32_t index, std::uint32_t generation) const;
 
-    std::array<float, 3> position(std::uint32_t index,
-                                  std::uint32_t generation) const;
+    std::array<double, 3> position(std::uint32_t index,
+                                   std::uint32_t generation) const;
     void set_position(std::uint32_t index, std::uint32_t generation,
-                      float x, float y, float z);
+                      double x, double y, double z);
 
-    std::array<float, 9> rotation(std::uint32_t index,
-                                  std::uint32_t generation) const;
+    std::array<double, 9> rotation(std::uint32_t index,
+                                   std::uint32_t generation) const;
     void set_rotation(std::uint32_t index, std::uint32_t generation,
-                      const std::array<float, 9>& r);
+                      const std::array<double, 9>& r);
 
-    std::array<float, 3> rotation_col(std::uint32_t index,
-                                      std::uint32_t generation,
-                                      int col) const;
+    std::array<double, 3> rotation_col(std::uint32_t index,
+                                       std::uint32_t generation,
+                                       int col) const;
 
     std::uint32_t live_count() const { return live_; }
     std::uint32_t capacity() const {
